+++ docs_src/src/changelog.md
@@ -4,6 +4,7 @@
 
 - Wielders being recruited are now announced. This is mainly for situations where they join via story events
 - Fixed the revealed buffer being reset after combat
+- Fixed resource generating buildings not appearing in the scanner. They have a dedicated "Resource generators" category
 
 ## V0.3.0